+++ src/pages/Contact/index.tsx
@@ -47,6 +47,7 @@ import { FaSuitcase, MdFilterList, MdSort, MdWeb } from 'react-icons/all';
 import MobileSortDialog from '../../components/MobileSortDialog';
 import MobileFilterDialog from '../../components/MobileFilterDialog';
 import { camelCase } from 'lodash';
+import WarhouseList from '../Account/Warehouse/WarhouseList';
 
 const ContactTypes = [
   {
@@ -108,6 +109,8 @@ export default function Contact(props) {
     isDelete: permissions[contactResource]?.isDelete
   });
   const [showEntityDialog, setShowEntityDialog] = useState(false);
+  const [openAddPlantsDialog, setOpenAddPlantsDialog] = useState(false);
+  const [isAddingWarehouse, setAddingWarehouse] = useState(false);
 
   const [filter, setFilter] = useState(queryType ? queryType : 'All Contacts');
   const [entities, setEntities] = useState([]);
@@ -958,6 +961,17 @@ export default function Contact(props) {
                         </MenuItem>
                       )}
                       {contactPermissions.isUpdate && (
+                      <MenuItem
+                        disabled={selectedRecords.length === 0 || [...new Set(selectedRecords.map((d) => d.accountNameId))].length > 1}
+                        onClick={() => {
+                          setOpenAddPlantsDialog(true)
+                          closeActions();
+                        }}
+                      >
+                        Add {routes.warehouse.title} &nbsp; <Chip size="small" label={selectedRecords.length} />
+                      </MenuItem>
+                    )}
+                      {contactPermissions.isUpdate && (
                         <MenuItem
                           disabled={selectedRecords.length === 0}
                           onClick={() => {
@@ -1141,6 +1155,31 @@ export default function Contact(props) {
               onOk={handleSingleDeleteContacts}
             />
           ) : null}
+          {openAddPlantsDialog &&
+          <WarhouseList
+            isCustomer={true}
+            api={`/customer-account/${selectedRecords[0].accountNameId}/warehouse`}
+            isAddingWarehouse={isAddingWarehouse}
+            addWarehouse={(selectedPlants:any) => {
+              setAddingWarehouse(true)
+              axiosInstance().post(`/customer-contact/assign-warehouse`, {
+                ids: selectedRecords.map((d:any) => d._id),
+                warehouse: selectedPlants.map((d:any) => d._id),
+              }).then(() => {
+                getContacts();
+                setAddingWarehouse(false)
+                setOpenAddPlantsDialog(false)
+              }).catch(err => {
+                toastConfig.setToastConfig(err)
+                setAddingWarehouse(false)
+                setOpenAddPlantsDialog(false)
+              })
+            }}
+            onClose={() => setOpenAddPlantsDialog(false) }
+            renderedFrom={renderedFrom}
+            assignedWarehouse={[]}
+          />
+        }
           {showEntityDialog ? (
             <EntitySelectionsDialog
               open={showEntityDialog}